refactor: renamed and reworked files in settingSelect folder.
